optimise ultimateadapter
